refactor: remove duplication for mocking fetch
