+++ src/order-book/api.spec.ts
@@ -1,15 +1,3 @@
-jest.mock('cross-fetch', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const fetchMock = require('jest-fetch-mock')
-  // Require the original module to not be mocked...
-  const originalFetch = jest.requireActual('cross-fetch')
-  return {
-    __esModule: true,
-    ...originalFetch,
-    default: fetchMock,
-  }
-})
-
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock'
 import { CowError } from '../common/types/cow-error'
 import { OrderBookApi } from './api'
