+++ src/trading/getPreSignTransaction.test.ts
@@ -1,16 +1,5 @@
 const GAS = '0x1e848' // 125000
 
-jest.mock('cross-fetch', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const fetchMock = require('jest-fetch-mock')
-  // Require the original module to not be mocked...
-  const originalFetch = jest.requireActual('cross-fetch')
-  return {
-    __esModule: true,
-    ...originalFetch,
-    default: fetchMock,
-  }
-})
 jest.mock('../common/generated', () => {
   const original = jest.requireActual('../common/generated')
