+++ src/trading/postCoWProtocolTrade.test.ts
@@ -1,15 +1,3 @@
-jest.mock('cross-fetch', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const fetchMock = require('jest-fetch-mock')
-  // Require the original module to not be mocked...
-  const originalFetch = jest.requireActual('cross-fetch')
-  return {
-    __esModule: true,
-    ...originalFetch,
-    default: fetchMock,
-  }
-})
-
 jest.mock('../order-signing', () => {
   return {
     OrderSigningUtils: {
