+++ src/trading/postSellNativeCurrencyOrder.test.ts
@@ -6,18 +6,6 @@ import { OrderBookApi, OrderKind } from '../order-book'
 import { postSellNativeCurrencyOrder } from './postSellNativeCurrencyOrder'
 import { WRAPPED_NATIVE_CURRENCIES } from '../common'
 
-jest.mock('cross-fetch', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const fetchMock = require('jest-fetch-mock')
-  // Require the original module to not be mocked...
-  const originalFetch = jest.requireActual('cross-fetch')
-  return {
-    __esModule: true,
-    ...originalFetch,
-    default: fetchMock,
-  }
-})
-
 jest.mock('../common/generated', () => {
   const original = jest.requireActual('../common/generated')
