+++ src/trading/tradingSdk.test.ts
@@ -1,14 +1,3 @@
-jest.mock('cross-fetch', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const fetchMock = require('jest-fetch-mock')
-  // Require the original module to not be mocked...
-  const originalFetch = jest.requireActual('cross-fetch')
-  return {
-    __esModule: true,
-    ...originalFetch,
-    default: fetchMock,
-  }
-})
 import { TradingSdk } from './tradingSdk'
 import { SupportedChainId } from '../common'
 import { TradeBaseParameters } from './types'
